ENH: Implement global delete as a specialization of an empty change
Add a:isDelete flag to ChangeGlobally#SetParameters().
Define duplicate delete mappings, with a default mapping to gX instead of gC.
FIX: Substitution to make all non-alphabetic delimiter characters and whitespace optional didn't correctly deal with newline \n and the escaped \/ and \\. Tweak the regexp to deal with those.
Avoid invoking SmartCase() on empty string. In the debugger, I've seen it turn it into a newline.